Add _form for create abnd edit template

+++ src/Controller/PinsController.php
@@ -31,7 +31,7 @@ class PinsController extends AbstractController
         $pin = new Pin;
         $form = $this->createForm(PinType::class, $pin);
 
-                    $form->handleRequest($request);
+                $form->handleRequest($request);
 
                 if ($form->isSubmitted() && $form->isValid()) {
                     $em->persist($pin);
@@ -80,10 +80,12 @@ class PinsController extends AbstractController
     /**
      * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"DELETE"})
      */
-    public function delete(Pin $pin, EntityManagerInterface $em): Response
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
     {
+        if ($this->isCsrfTokenValid('pin_deletion_'. $pin->getId(), $request->request->get('csrf_token') )) {
         $em->remove($pin);
         $em->flush();
+        }
 
         return $this->redirectToRoute('app_home');
     }
